0.46 release with FlatList fix

+++ Libraries/Image/RCTImageCache.m
@@ -1,8 +1,10 @@
 /**
  * Copyright (c) 2015-present, Facebook, Inc.
+ * All rights reserved.
  *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
  */
 
 #import "RCTImageCache.h"
@@ -14,7 +16,6 @@
 #import <React/RCTConvert.h>
 #import <React/RCTNetworking.h>
 #import <React/RCTUtils.h>
-#import <React/RCTResizeMode.h>
 
 #import "RCTImageUtils.h"
 
@@ -23,8 +24,8 @@ static const NSUInteger RCTMaxCachableDecodedImageSizeInBytes = 1048576; // 1MB
 static NSString *RCTCacheKeyForImage(NSString *imageTag, CGSize size, CGFloat scale,
                                      RCTResizeMode resizeMode, NSString *responseDate)
 {
-  return [NSString stringWithFormat:@"%@|%g|%g|%g|%lld|%@",
-          imageTag, size.width, size.height, scale, (long long)resizeMode, responseDate];
+    return [NSString stringWithFormat:@"%@|%g|%g|%g|%zd|%@",
+            imageTag, size.width, size.height, scale, resizeMode, responseDate];
 }
 
 @implementation RCTImageCache
@@ -37,7 +38,7 @@ static NSString *RCTCacheKeyForImage(NSString *imageTag, CGSize size, CGFloat sc
 {
   _decodedImageCache = [NSCache new];
   _decodedImageCache.totalCostLimit = 5 * 1024 * 1024; // 5MB
-  
+
   [[NSNotificationCenter defaultCenter] addObserver:self
                                            selector:@selector(clearCache)
                                                name:UIApplicationDidReceiveMemoryWarningNotification
@@ -46,7 +47,7 @@ static NSString *RCTCacheKeyForImage(NSString *imageTag, CGSize size, CGFloat sc
                                            selector:@selector(clearCache)
                                                name:UIApplicationWillResignActiveNotification
                                              object:nil];
-  
+
   return self;
 }
